backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
supabase==2.3.0
python-dotenv==1.0.0
httpx==0.25.2
redis==5.0.1
scikit-learn==1.3.2
xgboost==2.0.2
pandas==2.1.4
numpy==1.26.2
prophet==1.1.5
onnxruntime==1.16.3
shap==0.44.0
pyyaml==6.0.1
twilio==8.11.0
firebase-admin==6.3.0
geopy==2.4.1
shapely==2.0.2
